# openapi.yaml
openapi: 3.1.0
info:
  title: Scheduling App API
  version: 1.0.0
  description: |
    REST API for the Scheduling App — helps American Airlines flight attendants
    build optimized bidding schedules from their monthly bid sheet, enforcing
    rules from the 2024 AA/APFA Collective Bargaining Agreement (CBA).

    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require a valid
    Bearer token in the `Authorization` header.

    ## Error Envelope
    All error responses use a standard JSON body:
    ```json
    { "code": "string", "message": "string", "details": {} }
    ```

    ## Pagination
    List endpoints return `{ data: [...], page_state: "token|null" }`.
    Pass `page_state` as a query parameter to fetch the next page.

    ## Requirement Traceability
    Each operation includes an `x-req-ids` extension listing the requirement
    IDs from requirements.md that it fulfills.
servers:
  - url: https://api.schedulingapp.example.com/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /bid-periods:
    post:
      operationId: createBidPeriod
      tags: [Bid Periods]
      summary: Upload a bid sheet PDF and create a new bid period
      description: |
        Accepts a multipart upload containing the bid sheet PDF. The server
        begins asynchronous parsing and returns immediately with
        `parse_status: "processing"`. Poll the GET endpoint to check progress.
      x-req-ids: [REQ-001, REQ-002, REQ-022]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, name, effective_start, effective_end]
              properties:
                file:
                  type: string
                  format: binary
                  description: The bid sheet PDF file
                name:
                  type: string
                  description: "Human-readable name (e.g. 'January 2026')"
                effective_start:
                  type: string
                  format: date
                effective_end:
                  type: string
                  format: date
      responses:
        "202":
          description: Bid period created; parsing in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BidPeriod"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    get:
      operationId: listBidPeriods
      tags: [Bid Periods]
      summary: List all bid periods for the authenticated user
      x-req-ids: [REQ-019]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/PageState"
      responses:
        "200":
          description: Paginated list of bid periods
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BidPeriodList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /bid-periods/{bidPeriodId}:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"

    get:
      operationId: getBidPeriod
      tags: [Bid Periods]
      summary: Get a single bid period with parsing status
      x-req-ids: [REQ-019, REQ-022]
      responses:
        "200":
          description: Bid period details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BidPeriod"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteBidPeriod
      tags: [Bid Periods]
      summary: Delete a bid period and all associated data
      x-req-ids: [REQ-019]
      responses:
        "204":
          description: Deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /bid-periods/{bidPeriodId}/preferences:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"

    put:
      operationId: updateBidPeriodPreferences
      tags: [Bid Periods]
      summary: Set preference overrides for this bid period
      description: |
        Fields provided here override the user's defaults for this period only.
        Omitted fields inherit from the user's default preferences.
      x-req-ids: [REQ-009]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Preferences"
      responses:
        "200":
          description: Updated period preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Sequences ───────────────────────────────────────────────────────────────
...
  /bid-periods/{bidPeriodId}/bids:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"

    post:
      operationId: createBid
      tags: [Bids]
      summary: Create a new bid (draft) for this bid period
      x-req-ids: [REQ-011, REQ-013]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBidRequest"
      responses:
        "201":
          description: Bid created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    get:
      operationId: listBids
      tags: [Bids]
      summary: List all bids for this bid period
      x-req-ids: [REQ-019]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/PageState"
      responses:
        "200":
          description: List of bids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BidList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /bid-periods/{bidPeriodId}/bids/{bidId}:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"
      - $ref: "#/components/parameters/BidId"

    get:
      operationId: getBid
      tags: [Bids]
      summary: Get a bid with its full ranked entry list and summary
      x-req-ids: [REQ-011, REQ-016]
      responses:
        "200":
          description: Full bid detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateBid
      tags: [Bids]
      summary: Update bid entries (reorder, pin, exclude, rename, change status)
      x-req-ids: [REQ-013]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBidRequest"
      responses:
        "200":
          description: Updated bid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /bid-periods/{bidPeriodId}/bookmarks:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"

    post:
      operationId: createBookmark
      tags: [Bookmarks]
      summary: Bookmark a sequence
      x-req-ids: [REQ-021]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sequence_id]
              properties:
                sequence_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Bookmark created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"

    get:
      operationId: listBookmarks
      tags: [Bookmarks]
      summary: List all bookmarked sequences for this bid period
      x-req-ids: [REQ-021]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/PageState"
      responses:
        "200":
          description: Paginated bookmark list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /bid-periods/{bidPeriodId}/bookmarks/{bookmarkId}:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"
      - $ref: "#/components/parameters/BookmarkId"

    delete:
      operationId: deleteBookmark
      tags: [Bookmarks]
      summary: Remove a bookmark
      x-req-ids: [REQ-021]
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Filter Presets ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BidPeriodId:
      name: bidPeriodId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BidId:
      name: bidId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BookmarkId:
      name: bookmarkId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    PageState:
      name: page_state
      in: query
      schema:
        type: string
      description: Cursor token for the next page
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource already exists (e.g. duplicate email, duplicate bookmark)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: Request understood but cannot be processed (e.g. unparseable PDF)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples: ["VALIDATION_ERROR", "NOT_FOUND", "PARSE_FAILED"]
        message:
          type: string
          description: Human-readable error description
        details:
          type: object
          additionalProperties: true
          description: Additional context (field errors, etc.)

    # ── Auth ───────────────────────────────────────────────────────────────────
...
    Preferences:
      type: object
      description: All fields are optional; omitted fields retain their current value.
      properties:
        preferred_days_off:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 31
        preferred_layover_cities:
          type: array
          items:
            type: string
        avoided_layover_cities:
          type: array
          items:
            type: string
        tpay_min_minutes:
          type: integer
        tpay_max_minutes:
          type: integer
        preferred_equipment:
          type: array
          items:
            type: string
        report_earliest_minutes:
          type: integer
          description: Minutes from midnight
        report_latest_minutes:
          type: integer
        release_earliest_minutes:
          type: integer
        release_latest_minutes:
          type: integer
        avoid_redeyes:
          type: boolean
        prefer_turns:
          type: boolean
          nullable: true
        prefer_high_ops:
          type: boolean
          nullable: true
        cluster_trips:
          type: boolean
        prefer_holidays:
          type: boolean
          nullable: true
          description: "Prefer sequences operating on CBA §3.K holidays (100% premium)"
        prefer_speaker_sequences:
          type: boolean
          nullable: true
          description: "For language-qualified FAs: prefer Speaker sequences"
        weights:
          $ref: "#/components/schemas/PreferenceWeights"
    PreferenceWeights:
      type: object
      description: Relative importance (1-10) for each preference dimension
      properties:
        days_off:
          type: integer
          minimum: 1
          maximum: 10
        tpay:
          type: integer
          minimum: 1
          maximum: 10
        layover_city:
          type: integer
          minimum: 1
          maximum: 10
        equipment:
          type: integer
          minimum: 1
          maximum: 10
        report_time:
          type: integer
          minimum: 1
          maximum: 10
        release_time:
          type: integer
          minimum: 1
          maximum: 10
        redeye:
          type: integer
          minimum: 1
          maximum: 10
        trip_length:
          type: integer
          minimum: 1
          maximum: 10
        clustering:
          type: integer
          minimum: 1
          maximum: 10

    # ── Bid Period ─────────────────────────────────────────────────────────────
    BidPeriod:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        effective_start:
          type: string
          format: date
        effective_end:
          type: string
          format: date
        base_city:
          type: string
        source_filename:
          type: string
        parse_status:
          type: string
          enum: [pending, processing, completed, failed]
        parse_error:
          type: string
          nullable: true
        total_sequences:
          type: integer
        categories:
          type: array
          items:
            type: string
        issued_date:
          type: string
          format: date
          nullable: true
        preference_overrides:
          $ref: "#/components/schemas/Preferences"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BidPeriodList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BidPeriod"
        page_state:
          type: string
          nullable: true

    # ── Sequence ───────────────────────────────────────────────────────────────
...
    CommuteImpact:
      type: object
      description: "Per-sequence commute feasibility analysis based on the user's commute city"
      properties:
        first_day_feasible:
          type: boolean
          description: "Can the FA commute in same-day for the first report?"
        first_day_note:
          type: string
          description: "e.g. 'Report 14:26 — easy commute (earliest DCA→ORD arrives 08:30)'"
        last_day_feasible:
          type: boolean
          description: "Can the FA commute home same-day after the last release?"
        last_day_note:
          type: string
          description: "e.g. 'Release 12:07 — easy commute home'"
        hotel_nights_needed:
          type: integer
          minimum: 0
          maximum: 2
          description: "0, 1, or 2 based on first/last day feasibility"
        impact_level:
          type: string
          enum: [green, yellow, red]
          description: "green = easy, yellow = tight, red = hotel night needed"
...
    Bid:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bid_period_id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [draft, optimized, finalized, exported]
        entries:
          type: array
          items:
            $ref: "#/components/schemas/BidEntry"
        summary:
          $ref: "#/components/schemas/BidSummary"
        optimization_config:
          $ref: "#/components/schemas/OptimizationConfig"
        optimization_run_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OptimizationConfig:
      type: object
      nullable: true
      description: "Snapshot of inputs used for the last optimization run"
      properties:
        preferences_used:
          $ref: "#/components/schemas/Preferences"
        seniority_number:
          type: integer
        total_base_fas:
          type: integer
        pinned_ids:
          type: array
          items:
            type: string
            format: uuid
        excluded_ids:
          type: array
          items:
            type: string
            format: uuid
    BidEntry:
      type: object
      properties:
        rank:
          type: integer
        sequence_id:
          type: string
          format: uuid
        seq_number:
          type: integer
        is_pinned:
          type: boolean
        is_excluded:
          type: boolean
        rationale:
          type: string
          nullable: true
          description: "Plain-language explanation of why this sequence is at this rank"
        preference_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: "How well this sequence matches the user's weighted preferences (0.0–1.0)"
        attainability:
          type: string
          enum: [high, medium, low, unknown]
          description: "Estimated likelihood of being awarded given the user's seniority"
        date_conflict_group:
          type: string
          nullable: true
          description: "ID grouping sequences with overlapping operating dates; within a group only one can be awarded"
        commute_impact:
          $ref: "#/components/schemas/CommuteImpact"
          nullable: true
          description: "Commute feasibility for this sequence (present when user has commute_from set)"
    BidSummary:
      type: object
      properties:
        total_entries:
          type: integer
        total_tpay_minutes:
          type: integer
        total_block_minutes:
          type: integer
        total_tafb_minutes:
          type: integer
        total_days_off:
          type: integer
        sequence_count:
          type: integer
        leg_count:
          type: integer
        deadhead_count:
          type: integer
        international_count:
          type: integer
        domestic_count:
          type: integer
        layover_cities:
          type: array
          items:
            type: string
        date_coverage:
          $ref: "#/components/schemas/DateCoverage"
        conflict_groups:
          type: integer
          description: "Number of distinct date-conflict groups in the bid"
        total_credit_hours:
          type: number
          format: float
          description: "Estimated credited hours for Line of Time calculation"
        line_option:
          type: string
          enum: [standard, high, low]
        line_min_hours:
          type: integer
          description: "Minimum credit hours for selected line option (40, 70)"
        line_max_hours:
          type: integer
          description: "Maximum credit hours for selected line option (90, 110)"
        credit_hour_utilization:
          type: number
          format: float
          description: "credit_hours / line_max (0.0-1.0)"
        estimated_total_pay_cents:
          type: integer
          description: "Sum of estimated_pay_cents for realistic award projection"
        cba_violations:
          type: array
          items:
            type: string
          description: "CBA constraint violations detected (e.g., '7-day block limit exceeded on dates 12-18')"
        commute_warnings:
          type: array
          items:
            type: string
          description: "Commute-related warnings (e.g., 'Back-to-back trips SEQ 664→SEQ 512 with 14h gap — insufficient time to commute home to DCA')"
    DateCoverage:
      type: object
      description: "Tracks which dates in the bid period are/aren't covered by ranked sequences"
      properties:
        covered_dates:
          type: array
          items:
            type: integer
          description: "Days-of-month covered by at least one ranked sequence"
        uncovered_dates:
          type: array
          items:
            type: integer
          description: "Days-of-month with no ranked sequence — reserve risk"
        coverage_rate:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: "covered / total period days (1.0 = full coverage)"
    CreateBidRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        entries:
          type: array
          items:
            type: object
            required: [sequence_id]
            properties:
              sequence_id:
                type: string
                format: uuid
              is_pinned:
                type: boolean
                default: false
              is_excluded:
                type: boolean
                default: false
    UpdateBidRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [draft, finalized]
        entries:
          type: array
          items:
            type: object
            required: [sequence_id, rank]
            properties:
              sequence_id:
                type: string
                format: uuid
              rank:
                type: integer
              is_pinned:
                type: boolean
              is_excluded:
                type: boolean
    BidList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Bid"
        page_state:
          type: string
          nullable: true

    # ── Bookmark ───────────────────────────────────────────────────────────────
    Bookmark:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sequence_id:
          type: string
          format: uuid
        seq_number:
          type: integer
        created_at:
          type: string
          format: date-time
    BookmarkList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Bookmark"
        page_state:
          type: string
          nullable: true

    # ── Filter Preset ──────────────────────────────────────────────────────────
...